[refactor] 도메인 기반 구조 리팩토링
