+++ app/domain/recommend/recommend_router.py
@@ -1,7 +1,14 @@
 from fastapi import APIRouter, HTTPException
-from app.schemas.recommendation import RecommendRequest, FinalRecommendResponse, RecommendResponseData, ProblemRecommendation
+
+from app.domain.recommend.recommendation_schemas import (
+    FinalRecommendResponse,
+    ProblemRecommendation,
+    RecommendRequest,
+    RecommendResponseData,
+)
 
 router = APIRouter()
+
 
 @router.post("/", response_model=FinalRecommendResponse)
 async def get_recommendations(request: RecommendRequest):
@@ -11,22 +18,21 @@ async def get_recommendations(request: RecommendRequest):
         mock_recommendations = [
             ProblemRecommendation(
                 problem_id=1001,
-                reason_msg="[환경세팅] 기본 추천 문제입니다. 조건문 해석 능력을 키우기 좋습니다."
+                reason_msg="[환경세팅] 기본 추천 문제입니다. 조건문 해석 능력을 키우기 좋습니다.",
             ),
             ProblemRecommendation(
                 problem_id=1045,
-                reason_msg="[환경세팅] 유저 수준에 맞는 연습 문제입니다."
-            )
+                reason_msg="[환경세팅] 유저 수준에 맞는 연습 문제입니다.",
+            ),
         ]
 
         response_data = RecommendResponseData(
             user_id=request.user_id,
             scenario=request.scenario,
-            recommendations=mock_recommendations
+            recommendations=mock_recommendations,
         )
 
         return FinalRecommendResponse(data=response_data)
 
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
-
